Fixed some warning that showed up in gcc but not clang++, and vice versa (I guess it's useful to test under multiple compilers).
Bumped version # for next test release.


git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@3008 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/emucore/EventHandler.hxx
@@ -439,7 +439,7 @@ class EventHandler
         void saveMapping();
 
         const StellaJoystick* joy(int id) const {
-          return id < mySticks.size() ? mySticks[id] : NULL;
+          return id < (int)mySticks.size() ? mySticks[id] : NULL;
         }
 
       private:
